refactor: fix few newly introduced ruff-0.16 errors (#158)
* refactor: fix few newly introduced ruff-0.16 errors

* use more appropriate type for stateful system

ruff complains when class variables use some mutable data type. I
replace dict by Mapping

+++ matr1x/core/execthread.py
@@ -409,7 +409,7 @@ class ExecThread(threading.Thread):
                 break
 
         if initial_sleep_time > silent:
-            waited = int(round(time.time() - start_time))
+            waited = round(time.time() - start_time)
             text = f"Waited {waited} seconds"
             self.report(Message(text))
             self.report(Message(text, modifier=Modifier.TO_PROGRESS_LABEL))
